requirements.txt
asgiref
backports.zoneinfo
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
PyJWT
sqlparse
typing-extensions
